Add files via upload
## Extended Description

This project was created as part of my Java GUI practice using NetBeans. I wanted to build a simple form that collects student data and displays it interactively.

The application uses Java Swing and AWT to create a form with fields for name, age, and department. When the user clicks Submit, the data is displayed below using an event listener.

Through this project, I learned how to use layout managers, handle button events, and structure a GUI application in Java.

### Future Improvements
- Add input validation
- Save data to a file
- Improve layout using GridBagLayout